requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
azure-search-documents
azure-monitor-opentelemetry
azure-identity
azure-storage-blob
azure-ai-evaluation
openai
streamlit
python-dotenv
pytest
pypdf
ragas
scikit-learn
Pillow
httpx
pandas
mlflow<=3.5.0
azureml-mlflow
azure-ai-contentsafety
pytest
pytest-asyncio
kaggle
chromadb
python-json-logger


# langchain==0.1.0
# langchain-openai==0.0.2
# langchain-community==0.0.10
# openai==1.6.1

# # Azure Services
# azure-search-documents==11.4.0
# azure-ai-contentsafety==1.0.0
# azure-identity==1.15.0
# azure-monitor-opentelemetry==1.2.0
# azure-storage-blob==12.19.0

# # Document Processing
# pypdf==3.17.4
# # pandas==2.1.4
# pandas==1.5.3
# # numpy==1.24.4


# # Web Framework
# streamlit==1.29.0

# # MLflow & Monitoring
# mlflow<=3.5.0
# azureml-mlflow
# python-json-logger

# # Testing
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0

# # Utilities
# python-dotenv==1.0.0
# tqdm==4.66.1